Fix import generation for the D wrappers for C#.
The code was just importing the provided modules, not what was actually
be used by the code in the provided modules. So, it fell flat on its
face once it was given a modules with a function that used a type
imported by that module but not explicitly provided to autowrap. The
code now imports the modules for any types used in the signatures of the
functions being wrapped.