Fixed a couple of issues with the C# code type name generation.
Several instances where the return type name generated did not have the
fully qualified path in its name were fixed. Also, it was fixed so that
inside a type, the fully qualified name was not used, because in cases
where the type name matched the module name, it resulted in errors when
the C# code was compiled. The "simple" tests don't test any modules
inside of packages, so it's possible that there are further problems
related to how the names are generated which relate to packages which
have not been caught, but this at least improves the situation.