Fix hang in C# tests.
The D code generated to interact with C# was calling the code to attach
and detach the thread (including the thread-local, module constructors
and destructors) even when the thread was already attached, and unlike
rt_init and rt_term, those functions don't retain a ref-count; they
assume that you're only calling them when you need to. So, they were
running the constructors a second time and then running the destructors,
putting things in a weird state, and the result was that the GC would
hang when an attempt was made to allocate. The code now first checks
whether the thread has already been attached and skips doing any of that
work if it is attached. That fixed the hang, though since the current
tests all seem to be run on the same thread, some multi-threaded tests
will need to be added to better test to make sure that that code works
correctly.