Rework C# date/time support.
The way it was before was going to cause subtle problems. Now, there are
three different marshalling types for the date/time types - one for
Duration, one for Date, TimeOfDay, and DateTime, and one for SysTime.
Duration gets converted to TimeSpan. D's Date, TimeOfDay, and DateTime
get converted to C#'s DateTime with DateTimeKind.Utc. And SysTime gets
converted to DateTimeOffset.

For C#'s DateTime, the conversion between ticks and the year, month,
day, etc. properties is now done in C# so that what D considers those
properties to be will always match what C# considers them to be.